Onboarding game tweaks
